fix float

+++ HCf460/application/matlab/time_detection.c
@@ -145,12 +145,12 @@ int time_detection(FIFO_DataType data[], int data_size, int win_size_time, int
 						time_add_last = time_add;
 						Delayms(10);
 					}
-					if (time_times_rt != 0)
+					if (time_times_rt != (float)0)
 					{
 						mcu_dp_value_update(DPID_TIME_TIMES_RT, (int)((time_times_rt*100)+0.5f));
 						Delayms(10);
 					}
-					if (time_add_rt != 0)
+					if (time_add_rt != (float)0)
 					{
 						mcu_dp_value_update(DPID_TIME_ADD_RT, (int)((time_add_rt*100)+0.5f));
 					}
@@ -170,12 +170,12 @@ int time_detection(FIFO_DataType data[], int data_size, int win_size_time, int
 					time_add_last = time_add;
 					Delayms(10);
 				}
-				if (time_times_rt != 0)
+				if (time_times_rt != (float)0)
 				{
 					mcu_dp_value_update(DPID_TIME_TIMES_RT, (int)((time_times_rt*100)+0.5f));
 					Delayms(10);
 				}
-				if (time_add_rt != 0)
+				if (time_add_rt != (float)0)
 				{
 					mcu_dp_value_update(DPID_TIME_ADD_RT, (int)((time_add_rt*100)+0.5f));
 				}
